B0Map

Documentation for B0Map.

B0Map.AbstractGREMultiEcho

B0Map.FitOpt

B0Map.FitPar

B0Map.GREMultiEchoWF

B0Map.GREMultiEchoWF

B0Map.GREMultiEchoWFFW

B0Map.GREMultiEchoWFFW

B0Map.GREMultiEchoWFRW

B0Map.GREMultiEchoWFRW

B0Map.GSS

B0Map.GSS_fit_R2s

B0Map.GSS_fit_ϕ

B0Map.calc_A

B0Map.calc_Bb

B0Map.calc_par

B0Map.calc_par_chunk

B0Map.calc_∂Bb

B0Map.calc_∂Bb

B0Map.calc_∂f

B0Map.calc_∂∂Bb

B0Map.calc_∂∂Bb

B0Map.calc_∂∂f

B0Map.coil_sensitivities

B0Map.fatTrait

B0Map.fat_fraction

B0Map.fat_fraction

B0Map.fat_fraction

B0Map.fitOpt

B0Map.fitPar

B0Map.greMultiEchoWF

B0Map.greMultiEchoWFFW

B0Map.greMultiEchoWFRW

B0Map.local_fit

B0Map.local_fit_chunk

B0Map.set_num_phase_intervals

B0Map.update!

B0Map.update!

B0Map.update_A!

B0Map.update_fat_fraction!

B0Map.update_fat_fraction!

B0Map.ϕ_search_intervals

B0Map.∂Bb!

B0Map.∂Bb!

B0Map.∂∂Bb!

B0Map.∂∂Bb!

VP4Optim.A

VP4Optim.Bb!

VP4Optim.Bb!

VP4Optim.par_changed!

VP4Optim.par_changed!

VP4Optim.set_data!

VP4Optim.set_data!

VP4Optim.x_changed!

VP4Optim.x_changed!

VP4Optim.y!

VP4Optim.y!

VP4Optim.y!

VP4Optim.y_model

VP4Optim.∂Bb!

VP4Optim.∂Bb!

VP4Optim.∂∂Bb!

VP4Optim.∂∂Bb!

B0Map.AbstractGREMultiEcho — Type

AbstractGREMultiEcho{Ny,Nx,Nc,T}

Abstract supertype of multi-echo GRE sequences

Type parameters

Ny::Int: Number of acquired data == number(echoes) * number(coils)

Nx::Int: Number of variable parameters x ⊆ {ϕ, R2s, ...} (relevant for optimization)

Nc::Int: Number of linear coefficients (e.g. number of coil elements)

T::Union{Float64, ComplexF64}: acquired data type

B0Map.FitOpt — Type

Data structure holding the fit parameters.

Local Fit

n_ϕ::Int: number of golden section search (GSS) intervals for the initial phase search (with R2s == 0).

ϕ_rngs::Array: GSS intervals, associated with n_ϕ.

R2s_rng::Array: Search range for R2s.

ϕ_acc::Float64: Required GSS accuracy for ϕ_acc

R2s_acc::Float64: Required GSS accuracy for R2s_acc

optim::Bool: Non-linear optimiztion in addition to GSS? (Requires gradients to be implemented for the GRE model.)

General

n_chunks::Int: Number of chunks to profit from multi-threaded execution.

Remark

Use ϕsearchintervals to modify n_ϕ, since only then ϕ_rngs will be set properly.

B0Map.FitPar — Type

Data structure for fitting images.

Content

Specifics of the GRE acquisition and signal/tissue model

Mask to specify ROI

Arrays for the fitted internal parameters ϕ and R2s

Arrays for the resulting linear coil-dependent coefficients c(ϕ, R2s)

Goodness of fit

Scope

Local and non-local (regularized) fitting

B0Map.GREMultiEchoWF — Type

VP4Optim model

Scope

RF spoiled multi-echo GRE sequence

Water-fat tissue model

Specifics

Water and fat are constrained to have equal phase at zero echo time.

Support for multiple coil elements and coil noise covariance matrix.

The fat fraction can either be calculated to minimize the cost function $\chi^2$ for fixed ϕ and R2s (this is the default setting) or set manually.

B0Map.GREMultiEchoWF — Method

GREMultiEchoWF(::Val{Ny}, ::Val{Nx}, ::Val{Nc}, ::Val{Nt}, ts, B0, ppm_fat, ampl_fat, 
    precession, x_sym, Δt, mode, cov_mat) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.GREMultiEchoWFFW — Type

VP4Optim model

Scope

RF spoiled multi-echo GRE sequence

Water-fat tissue model

Specifics

Water and fat are not constrained to have equal phase at zero echo time.

Single coil only

B0Map.GREMultiEchoWFFW — Method

GREMultiEchoWFFW(::Val{Ny}, ::Val{Nx}, ts, B0, ppm_fat, ampl_fat, 
    precession, x_sym, Δt) where {Ny,Nx}

Auxiliary function

B0Map.GREMultiEchoWFRW — Type

VP4Optim model

Scope

RF spoiled multi-echo GRE sequence

Water-fat tissue model

Specifics

Water and fat components are described by real weights, multiplied with a common phase factor.

Single coil only.

No partial derivatives implemented

B0Map.GREMultiEchoWFRW — Method

GREMultiEchoWFRW(::Val{Ny}, ::Val{Nx}, ::Val{Nc}, ::Val{Nt}, ts, B0, ppm_fat, ampl_fat,
precession, x_sym, Δt) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.GSS — Method

GSS(fun::Function, var_rng, acc; show_all=false)

Search for minimum of given function fun with golden section search (GSS).

Arguments

fun::Function: real valued function of a single real valued argument.

var_rng::T <: Union{Vector, Tuple}: search interval boundaries, length(var_rng) == 2

acc::Float64: Required accuracy of location.

show_all::Bool: Also return visited locations and their values (default: false)

Output

show_all ? ((x_opt, fun(x_opt)), xs, fs) : (x_opt, fun(x_opt)), where

x_opt is the location of the minimum with value fun(x_opt)

xs::Vector{Float64} is the vector of all tested locations and fs = map(x -> fun(x), xs)

B0Map.GSS_fit_R2s — Method

GSS_fit_R2s(gre::AbstractGREMultiEcho, ϕ, R2s_rng; R2s_acc=1e-4)

Calculate the best R2s for given ϕ with GSS Auxiliary function

B0Map.GSS_fit_ϕ — Method

GSS_fit_ϕ(gre::AbstractGREMultiEcho, R2s, ϕ_rng; ϕ_acc=1e-4)

Auxiliary function

B0Map.calc_A — Method

calc_A(gre::GREMultiEchoWFFW{Ny,Nx}) where {Ny,Nx}

Auxiliary method

B0Map.calc_Bb — Method

calc_Bb(gre::GREMultiEchoWFFW, A)

Auxiliary method

B0Map.calc_par — Method

calc_par(fitpar::FitPar{T}, parfun::Function, res::AbstractArray, n_chunks=8Threads.nthreads()) where {T<:AbstractGREMultiEcho}

Extract model specific information and store it in array res.

Arguments

fitpar::FitPar: Fit parameters

fitopt::FitOpt: Fit options

parfun::Function: Used to call parfun(gre::T) for each location in fitpar.S, based upon the local data and fit estimates.

res::AbstractArray: Allocated space for the results (size(res) == size(fitpar.S)).

B0Map.calc_par_chunk — Method

calc_par_chunk(gre::AbstractGREMultiEcho, fitpar::FitPar, parfun::Function, res::AbstractArray, cis_chunk)

Auxiliary function

B0Map.calc_∂Bb — Method

calc_∂Bb(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}, sy::SVector{Nsy}) where {Ny,Nx,Nc,Nt,Nsy}

Auxiliary routine

B0Map.calc_∂Bb — Method

calc_∂Bb(gre::GREMultiEchoWFFW{Ny,Nx}, A) where {Ny,Nx}

Auxiliary method

B0Map.calc_∂f — Method

calc_∂f(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.calc_∂∂Bb — Method

calc_∂∂Bb(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}, sy::SVector{Nsy}) where {Ny,Nx,Nc,Nt,Nsy}

Auxiliary routine

B0Map.calc_∂∂Bb — Method

calc_∂∂Bb(gre::GREMultiEchoWFFW{Ny,Nx}, tA) where {Ny,Nx}

Auxiliary method

B0Map.calc_∂∂f — Method

calc_∂∂f(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.coil_sensitivities — Method

coil_sensitivities(gre::GREMultiEchoWF)

Calculates and returns the coil sensitivities.

Remarks

If the coil noise covariance matrix $\Psi$ differs from the unit matrix, the routine VP4Optim.c does not return the actual coil sensitivities.

Otherwise it does.

B0Map.fatTrait — Method

fatTrait(gre::GREMultiEchoWF)

Auxiliary routine

B0Map.fat_fraction — Method

fat_fraction(gre::GREMultiEchoWFFW)

Calculates and returns fat fraction.

B0Map.fat_fraction — Method

fat_fraction(gre::GREMultiEchoWFRW)

Returns $|r_f| / (|r_w| + |r_f|)$

B0Map.fat_fraction — Method

fat_fraction(gre::GREMultiEchoWF)

Just returns the actual value of f, whether calculated automatically or set manually.

B0Map.fitOpt — Method

fitOpt(gre::AbstractGREMultiEcho)

Default constructor for FitOpt

Arguments

gre::AbstractGREMultiEcho: Initialized structure with GRE sequence parameters and signal/tissue model.

Default values

n_ϕ == 3

ϕ_rngs == ϕ_search_intervals(n_ϕ, gre.ϕ_scale)

R2s_rng == [0.0, 1.0]

ϕ_acc == 1.e-4

R2s_acc == 1.e-4

optim == true

n_chunks == 8Threads.nthreads()

Remarks

gre is only needed, if Δt != ΔTE (to define an effective bandwidth in case of non-equidistant echo times)

B0Map.fitPar — Method

fitPar(gre::AbstractGREMultiEcho{Ny,Nx,Nc,T}, data::AbstractArray, S::AbstractArray) where {Ny,Nx,Nc,T}

Constructor for FitPar

Arguments

gre::AbstractGREMultiEcho: Initialized structure with GRE sequence parameters and signal/tissue model.

data::AbstractArray: Complex multi-echo (if available also multi-coil) GRE data for each location.

S::AbstractArray: Mask to specify ROI.

Remarks

size(S) defines size and spatial dimensions (typically 2 or 3) of the data block.

Mandatory: size(data)[ndims(S)] == size(S)

The format of data must be such that VP4Optim.setdata!(gre, reshape(data, size(S)..., :)[ci,:]) works for any ci ∈ CartesianIndices(S)

For single-coil images, this means ndims(data) == ndims(S) + 1 with the last index enumerating the echoes.

For multi-coil data, we typically have ndims(data) == ndims(S) + 2 with the last but one index enumerating echo times and the last index enumerating the coils.

B0Map.greMultiEchoWF — Function

greMultiEchoWF(ts, B0, ppm_fat, ampl_fat, precession, mode=:auto_fat; 
    x_sym=nothing, Δt=nothing, n_coils=1, cov_mat=ones(ComplexF64,1,1))

Constructor

Arguments

ts::Vector{Float64}: Echo times [ms]

B0::Float64: Main field strength [T]

ppm_fat::Vector{Float64}: Chemical shift of fat peaks

ampl_fat::Vector{Float64}: relative fat peak amplitudes (all positive with sum(ampl_fat) ≈ 1)

precession::Symbol: Orientation of precession ∈ {:clockwise, :counterclockwise}

mode::Symbol: By default (mode == :auto_fat) the fat fraction f is calculated automatically. Alternatively (mode == :manual_fat), it can also be set manually.

x_sym::Vector{Symbol}: Vector of variable parameters ∈ {:ϕ, :R2s}, default: [:ϕ, :R2s]

Δt::Union{Float64, Nothing}: Allows to adjust the frequency bandwidth $2π/Δt$ in case of non-equidistant echoes. Default: Δt equals the average echo spacing.

n_coils::Integer: Number of coil elements (default == 1). Note: n_coils == Nc

cov_mat::AbstractMatrix: Coil noise covariance matrix, if available. Default: Nc x Nc unit matrix

B0Map.greMultiEchoWFFW — Method

greMultiEchoWFFW(ts, B0, ppm_fat, ampl_fat, precession; x_sym=[:ϕ, :R2s], Δt=nothing)

Constructor

Arguments

ts::Vector{Float64}: Echo times [ms]

B0::Float64: Main field strength [T]

ppm_fat::Vector{Float64}: Chemical shift of fat peaks

ampl_fat::Vector{Float64}: relative fat peak amplitudes (all positive with sum(ampl_fat) ≈ 1)

precession::Symbol: Orientation of precession ∈ {:clockwise, :counterclockwise}

x_sym::Vector{Symbol}: Vector of variable parameters ∈ {:ϕ, :R2s}, default: [:ϕ, :R2s]

Δt::Union{Float64, Nothing}: Allows to adjust the frequency bandwidth $2π/Δt$ in case of non-equidistant echoes. Default: Δt equals the average echo spacing.

B0Map.greMultiEchoWFRW — Method

greMultiEchoWFRW(ts, B0, ppm_fat, ampl_fat, precession; x_sym=[:ϕ, :R2s], Δt=nothing)

Constructor

Arguments

ts::Vector{Float64}: Echo times [ms]

B0::Float64: Main field strength [T]

ppm_fat::Vector{Float64}: Chemical shift of fat peaks

ampl_fat::Vector{Float64}: relative fat peak amplitudes (all positive with sum(ampl_fat) ≈ 1)

precession::Symbol: Orientation of precession ∈ {:clockwise, :counterclockwise}

mode::Symbol: By default (mode == :auto_fat) the fat fraction f is calculated automatically. Alternatively (mode == :manual_fat), it can also be set manually.

x_sym::Vector{Symbol}: Vector of variable parameters ∈ {:ϕ, :R2s}, default: [:ϕ, :R2s]

Δt::Union{Float64, Nothing}: Allows to adjust the frequency bandwidth $2π/Δt$ in case of non-equidistant echoes. Default: Δt equals the average echo spacing.

B0Map.local_fit — Method

local_fit(fitpar::FitPar{T}, fitopt::FitOpt) where {T<:AbstractGREMultiEcho}

Fit data to multi-echo GRE model locally.

Arguments

fitpar::FitPar: Fit parameters (see FitPar and fitPar](@ref fitPar)).

fitopt::FitOpt: Fit options (see FitOpt and fitOpt](@ref fitOpt)).

Remarks

Fits fitpar.data to all points specified by fitpar.S.

Makes use of multi-threading. Just start julia with as many workers as you like to use.

If isempty(fitopt.ϕ_rngs) == true, the GSS fit is only with respect to R2s, based upon the local phase, specified in fitpar.ϕ.

Otherwise, for each interval in fitopt.ϕ_rngs a GSS search with respect to ϕ is performed, each time for R2s == 0. For each of these phases fixed, a subsequent GSS search with respect to R2s is performed. The combination [ϕ, R2s], which produces the best fit is selected.

Optionally (fitopt.optim == true), for the best GSS estimate [ϕ, R2s] is refined with a nonlinear fit.

The final estimates [ϕ, R2s] are stored in fitpar, together with the linear coefficients VP4Optim.c and the goodness of fit VP4Optim.χ2.

B0Map.local_fit_chunk — Method

local_fit_chunk(gre, fitpar, fitopt, cis_chunk)

Auxiliary function

B0Map.set_num_phase_intervals — Method

set_num_phase_intervals(fitpar, fitopt, n_ϕ)

Sets the GSS intervals for the initial ϕ search (with R2s == 0.0)

Arguments

fitpar::FitPar: Fit parameters

fitopt::FitOpt: Fit options

n_ϕ: Number of search intervals

Remarks

Resets field n_ϕ in fitopt and then calls ϕsearchintervals.

B0Map.update! — Method

update!(gre::GREMultiEchoWFRW{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.update! — Method

update!(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.update_A! — Method

update_A!(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.update_fat_fraction! — Method

update_fat_fraction!(::AutoFat, gre::GREMultiEchoWF)

Auxiliary routine

B0Map.update_fat_fraction! — Method

update_fat_fraction!(::ManualFat, ::GREMultiEchoWF)

Auxiliary routine

B0Map.ϕ_search_intervals — Method

ϕ_search_intervals(n_ϕ, ϕ_scale=1.0)

Calulates the GSS intervals for the initial ϕ search (with R2s == 0.0)

Arguments

gre::AbstractGREMultiEcho: Initialized structure with GRE sequence parameters and signal/tissue model.

n_ϕ::Int: number of golden section search (GSS) intervals

Remarks

Background: For non-equidistant echo times, the 2π-periodicity with respect to ϕ no longer holds.

gre contains a field ϕ_scale == Δt / ΔTE to define an effective periodicity (or better search range) ϕ_scale * 2π via the optional parameter Δt.

Should not be called directly. Use setnumphase_intervals instead.

B0Map.∂Bb! — Method

∂Bb!(::AutoFat, gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.∂Bb! — Method

∂Bb!(::ManualFat, gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.∂∂Bb! — Method

∂∂Bb!(::AutoFat, gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.∂∂Bb! — Method

∂∂Bb!(::ManualFat, gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

VP4Optim.A — Method

VP.A(gre::GREMultiEchoWFFW)

Method described in VP4Optim

VP4Optim.Bb! — Method

VP.Bb!(gre::GREMultiEchoWFFW)

Method described in VP4Optim

VP4Optim.Bb! — Method

VP.Bb!(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Method described in VP4Optim

VP4Optim.par_changed! — Method

VP.par_changed!(gre::GREMultiEchoWFRW)

Method described in VP4Optim TBW

VP4Optim.par_changed! — Method

VP.par_changed!(gre::GREMultiEchoWF)

Method described in VP4Optim TBW

VP4Optim.set_data! — Method

VP.set_data!(gre::GREMultiEchoWFRW{Ny,Nx,Nc,Nt}, data::AbstractArray) where {Ny,Nx,Nc,Nt}

Stores data

Arguments

data::AbstractArray: Complex data as a vector of length Ny/2.

Remarks

Input is transformed into a real vector and handed over to VP4Optim.y!.

VP4Optim.set_data! — Method

VP.set_data!(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}, data::AbstractArray) where {Ny,Nx,Nc,Nt}

Sets new data and transforms them, if necessary.

Arguments

data::AbstractArray: Complex data for all time points and coils. The elements in data must be ordered such that the expression SMatrix{Nt,Nc}(data) produces the correct data matrix $y_{j\gamma}$.

Remarks

The data are transformed based upon the coil noise covariance matrix, as described in the documentation.

This is followed by a call of VP4Optim.y! with the transformed data as arguments.

VP4Optim.x_changed! — Method

VP.x_changed!(gre::GREMultiEchoWFRW)

Method described in VP4Optim

VP4Optim.x_changed! — Method

VP.x_changed!(gre::GREMultiEchoWF)

Method described in VP4Optim

VP4Optim.y! — Method

VP.y!(gre::GREMultiEchoWFRW{Ny,Nx,Nc,Nt}, new_y::AbstractArray) where {Ny,Nx,Nc,Nt}

Method described in VP4Optim

VP4Optim.y! — Method

VP.y!(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}, new_y::AbstractArray) where {Ny,Nx,Nc,Nt}

Method described in VP4Optim

VP4Optim.y! — Method

VP.y!(gre::GREMultiEchoWFFW{Ny,Nx}, new_y::AbstractArray) where {Ny,Nx}

Method described in VP4Optim

VP4Optim.y_model — Method

VP.y_model(gre::GREMultiEchoWF)

Method described in VP4Optim

VP4Optim.∂Bb! — Method

VP.∂Bb!(gre::GREMultiEchoWFFW)

Method described in VP4Optim

VP4Optim.∂Bb! — Method

VP.∂Bb!(gre::GREMultiEchoWF)

Method described in VP4Optim

VP4Optim.∂∂Bb! — Method

VP.∂∂Bb!(gre::GREMultiEchoWF)

Method described in VP4Optim

VP4Optim.∂∂Bb! — Method

VP.∂∂Bb!(gre::GREMultiEchoWFFW{Ny,Nx}) where {Ny,Nx}

Method described in VP4Optim