Constrained water-fat mixture

See article for a more in-depth description.

Description

We consider a set of Ny = Nt * Nc complex local RF-spoiled multi-echo GRE signals $y_{j\gamma}$ acquired at Nt different echo times $t_j$ in Nc ($\ge 1$) coil elements (indexed by $\gamma$). Assuming a water-fat mixture, the signal model is set up as follows

\[ \hat{y}_{j\gamma} \; := \; c_\gamma \left[\, 1 + f\,\left(\beta_j - 1\right)\,\right]\,e_j \qquad\qquad e_{j} \; := \; e^{i\,\omega\,t_{j}}\,e^{-R_2^{\ast}\,t_{j}}\]

with $0\le f \le 1$ and the multi-peak fat model given by

\[ \beta_{j} \; := \;\sum_{k=1}^{m}\,\alpha_{k}\,e^{i\,\omega_{k}t_{j}} \qquad\qquad \sum_{k=1}^m\,\alpha_k \;= \;1\]

The real-valued amplitudes $\alpha_k>0$ and chemical shifts $\omega_k$ are fixed and assumed to be known.

Note

This model guarantees the water peak and all fat peaks to have equal phase at $t=0$. One can recover the Unconstrained water-fat mixture model via

\[c_{w\gamma} = c_\gamma\left(1-f\right) \qquad\qquad c_{f\gamma} = c_\gamma\,f\]

After regrouping these equations

\[c_{\gamma} = c_{w\gamma} + c_{f\gamma} \qquad\qquad f = \frac{\left|c_{f\gamma}\right|}{\left|c_{f\gamma}\right| + \left|c_{w\gamma}\right|}\]

it becomes evident that $f$ is just the relative fat fraction and $c_\gamma$ an in-phase signal, weighted by the coil sensitivity.

Due to finite coil overlap, data noise $\eta_{j\gamma} = y_{j\gamma} - \hat{y}_{j\gamma}$ can be correlated, which is then expressed by a Hermitian coil noise covariance matrix $\bm{\Psi}$

\[ \left\langle \eta_{j,\gamma}\, \eta^\ast_{j^\prime\gamma^\prime}\right\rangle =: \Psi_{\gamma\gamma^\prime}\cdot \delta_{jj^\prime} = \sum_k\,\sigma^2_k\cdot U_{\gamma k}U^*_{\gamma^\prime k}\]

Based upon the eigendecomposition in the last equation, the transformations $c_\gamma \to c_k$ and $y_{j\gamma} \to y_{jk}$

\[ c_k := \sum_\gamma\,\frac{1}{\sigma_k}\,U^*_{\gamma k}\,c_\gamma \qquad\qquad y_{jk} := \sum_\gamma\,\frac{1}{\sigma_k}\,U^*_{\gamma k}\,y_{j\gamma}\]

allow to write the multi-coil ML cost function $\chi^2$ in the conventional form as expected by VARPRO. (See [link to follow] for more details.)

In summary, the variable parameters of this model are therefore given by $c_k \in \mathbb{C}$ and $\omega, R_2^\ast, f\in \mathbb{R}$.

Writing the signal model in vector notation $\bm{y}=\bm{A}\bm{c}$, leads to the following definition of the Ny $\times$ Nc matrix $\bm{A}$

\[ A_{jk} \equiv \left[\, 1 + f\,\left(\beta_j - 1\right)\,\right]\,e_j \qquad\qquad \forall \,k\]

and the linear coefficient vector $[\bm{c}]_k= c_k$.

Note

Note that the incorporating $f$ into $\bm{A}$ increases the complexity of the VARPRO cost function $\chi^2$. As shown in [link to follow], it is possible to determine $f=f\left(\omega,R_2^\ast\right)$ efficiently and thereby remove its explicit appearance in $\chi^2$. This correponds to the default setting mode = :autofat in the constructor below.

In terms of variable projection, specifically the notation $\bm{A}\left(\bm{x}\right)$ introduced in VP4Optim, the elements in the variable vector $\bm{x}$ are therefore given by a (sub)set $\subseteq \left\{\omega, R_2^\ast\right\}$.

Type

B0Map.GREMultiEchoWF — Type

VP4Optim model

Scope

RF spoiled multi-echo GRE sequence

Water-fat tissue model

Specifics

Water and fat are constrained to have equal phase at zero echo time.

Support for multiple coil elements and coil noise covariance matrix.

The fat fraction can either be calculated to minimize the cost function $\chi^2$ for fixed ϕ and R2s (this is the default setting) or set manually.

Constructor parameters

The parameters for the constructor are defined in

B0Map.ModParWF — Type

ModParWF <: VP4Optim.ModPar{GREMultiEchoWF}

Parameters to setup an instance of GREMultiEchoWF

Fields

ts::Vector{Float64}: Echo times [ms]

B0::Float64: Field strength [T]

ppm_fat::Vector{Float64}: Chemical shift of fat peaks [ppm]

ampl_fat::Vector{Float64}: Relative amplitudes of fat peaks (≥ 0, add up to one)

precession::Symbol: Direction of precession ∈ (:clockwise, :counterclockwise)

n_coils::Int: Number of coils

cov_mat::Matrix{ComplexF64}: Coil covariance matrix

mode::Symbol: Calculate fat (mode == :auto_fat) or explicitly set it (mode == :manual_fat)

x_sym::Vector{Symbol}: Variable parameters

Δt::Float64: Effective echo spacing (see docs), Δt == 0 means Δt = mean(ΔTE)

It can be set up with VP4Optim.modpar.

Constructor

B0Map.GREMultiEchoWF — Method

GREMultiEchoWF(pars::ModParWF)

Constructor

Arguments

pars::ModParWF: Model parameters to instantiate the model. See ModParWF.

Support for VP4Optim

All methods are fully implemented.

Partial derivatives up to second order are provided.

Specific API

Note

For $\Psi \ne \mathrm{I}$, data must be (re)set with VP.set_data!(gre, data) instead of VP.y!(gre, data).

The data argument must be ordered such that the expression SMatrix{Nt,Nc}(data) produces the correct data matrix $y_{j\gamma}$.

Note

The fat fraction is simply given by $f(\omega, R_2^\ast)$ and returned by fat_fraction.

B0Map.coil_sensitivities — Function

coil_sensitivities(gre::GREMultiEchoWF)

Calculate and return the coil sensitivities.

Remarks

If the coil noise covariance matrix $\Psi$ differs from the unit matrix $\mathrm{I}$, the linear VARPRO coefficients, returned by the routine VP4Optim.c, are not equal to the actual coil sensitivities.

Otherwise they are.