Unconstrained water-fat mixture

Description

We consider a set of Ny complex local signals $y_j$ acquired at different echo times $t_j$ in an RF-spoiled multi-echo GRE sequence. A water-fat signal model is assumed in the usual manner

\[ \hat{y}_{j} \; := \; \bigl(c_{w} + c_{f}\,\beta_j\bigr)\,e_j \qquad\qquad e_{j} \; := \; e^{i\,\omega\,t_{j}}\,e^{-R_2^{\ast}\,t_{j}}\]

with the multi-peak fat model given by

\[ \beta_{j} \; := \;\sum_{k=1}^{m}\,\alpha_{k}\,e^{i\,\omega_{k}t_{j}} \qquad\qquad \sum_{k=1}^m\,\alpha_k \;= \;1\]

The real-valued amplitudes $\alpha_k>0$ and chemical shifts $\omega_k$ are fixed and assumed to be known.

The variable parameters of this model are therefore given by $c_w,c_f \in \mathbb{C}$ and $\omega, R_2^\ast\in \mathbb{R}$.

Writing the signal model in vector notation $\bm{y}=\bm{A}\bm{c}$, leads to the following definition of the Ny $\times$ Nc matrix $\bm{A}$

\[A_{j1} \;=\; e_j \qquad\text{and}\qquad A_{j2} \;:=\; \beta_j\, e_j\]

together with the linear coefficient vector $\bm{c} = \left[\,c_w\,c_f\,\right]^T$.

In terms of variable projection, specifically the notation $\bm{A}\left(\bm{x}\right)$ introduced in VP4Optim, the elements in the variable vector $\bm{x}$ are given by a (sub)set $\subseteq \left\{\omega, R_2^\ast\right\}$.

Note

In VARPRO, the coefficients $c_w$ and $c_f$ are calculated in closed form (cf. the intro section in VP4Optim). This means that there is not way to enforce them to have equal phase. But since all $\alpha_k$ are assumed to be real and positive, this would be a consistency requirement. For alternative models, see

Real-valued water-fat mixture: Aligns water and fat, but allows for opposite direction

Constrained water-fat mixture: Full implementation of the constraint

Type

B0Map.GREMultiEchoWFFW — Type

VP4Optim model

Scope

RF spoiled multi-echo GRE sequence

Water-fat tissue model

Specifics

Water and fat are not constrained to have equal phase at zero echo time.

Single coil only

Constructor parameters

The parameters for the constructor are defined in

B0Map.ModParWFFW — Type

ModParWFFW <: VP4Optim.ModPar{GREMultiEchoWFFW}

Parameters to setup an instance of GREMultiEchoWF

Fields

ts::Vector{Float64}: Echo times [ms]

B0::Float64: Field strength [T]

ppm_fat::Vector{Float64}: Chemical shift of fat peaks [ppm]

ampl_fat::Vector{Float64}: Relative amplitudes of fat peaks (≥ 0, add up to one)

precession::Symbol: Direction of precession ∈ (:clockwise, :counterclockwise)

x_sym::Vector{Symbol}: Variable parameters

Δt::Float64: Effective echo spacing (see docs), Δt == 0 means Δt = mean(ΔTE)

It can be set up with VP4Optim.modpar.

Constructor

B0Map.GREMultiEchoWFFW — Method

GREMultiEchoWFFW(pars::ModParWFFW)

Constructor

Arguments

pars::ModParWFFW: Model parameters to instantiate the model. See ModParWFFW.

Support for VP4Optim

All methods are fully implemented.

Partial derivatives up to second order are provided.

Specific API

Note

The fat fraction is defined as

\[\frac{\left|c_{f}\right|}{\left|c_w\right| + \left|c_f\right|}\]

and returned by fat_fraction.