Set up parameters

Sequence and tissue

Any available information about the acquisition and tissue, including the data and allocated arrays for the fit results is stored in a single data structure

B0Map.FitPar — Type

Data structure for fitting images.

Content

grePar::VP4Optim.ModPar{AbstractGREMultiEcho}: Specifics of the GRE acquisition and signal/tissue model

data::Array: Complex data

S::Array: Mask to specify ROI

ϕ::Array: Phase $\phi = \omega \cdot \Delta t$ (fit parameter)

R2s::Array: Relaxation rate $R_2^\ast$ (fit parameter)

c::Array: Array of linear VARPRO coefficient vectors $\bm{c}(\phi, R_2^\ast)$

χ2::Array: Least squares residual $\chi^2$

Scope

Local and regularized fit

To create an instance, we just call a constructor

B0Map.fitPar — Function

fitPar(grePar::VP.ModPar{T}, data::AbstractArray, S::AbstractArray) where {T <: AbstractGREMultiEcho}

Constructor for FitPar

Arguments

grePar::VP4Optim.ModPar{AbstractGREMultiEcho}: Specifics of the GRE acquisition and signal/tissue model

data::AbstractArray: Complex multi-echo (if available also multi-coil) GRE data for each location.

S::AbstractArray: Mask to specify ROI.

Remarks

size(S) defines size and spatial dimensions (typically 2 or 3) of the data block.

Mandatory: size(data)[1:ndims(S)] == size(S)

The format of data must be such that VP4Optim.setdata!(gre, reshape(data, size(S)..., :)[ci,:]) works for any ci ∈ CartesianIndices(S)

For single-coil images, this means ndims(data) == ndims(S) + 1 with the last index enumerating the echoes.

For multi-coil data, we typically have ndims(data) == ndims(S) + 2 with the last but one index enumerating echo times and the last index enumerating the coils.

Example

import VP4Optim as VP
import B0Map as BM

# Mandatory acquisition parameters
nTE = 6                             # number of echoes
t0 = 0.5                            # first echo time [ms]
ΔTE = 1.0                           # echo spacing [ms]
TEs = [range(t0, t0 + (nTE-1) * ΔTE, nTE);]  # resulting echo times
B0 = 3.0                            # scanner field strength [T]
precession = :counterclockwise      # orientation of precession (depends on the scanner)

# Optional parameters
n_coils = 6                         # number of coils (if reconstructed separately)
cov_mat = get_cov_mat(...)          # coil covariance matrix (if available)

# Define water-fat tissue model (we only need to specify the fat component)
ppm_fat = [-3.80, -3.40, -2.60, -1.94, -0.39, 0.60]
ampl_fat = [0.087, 0.693, 0.128, 0.004, 0.039, 0.048]

# set up model constructor parameters
grePar = VP.modpar(BM.ModParWF;
    ts = TEs,                       # required
    B0 = B0,                        # required
    ppm_fat = ppm_fat,              # required
    ampl_fat = ampl_fat,            # required
    precession = precession,        # required
    n_coils = n_coils,              # optional (if ≠ 1)
    cov_mat = cov_mat,              # optional (if of relevance)
    )
    
# Prepare data and ROI
# (For the required dimensions, see the docs of the constructor fitPar().)
data = get_from_somewhere(...)
S = define_ROI_somehow(data, ...)

# create an instance of FitPar
fitpar = BM.fitPar(grePar, data, S) 

Fit options

All options, affecting the fit routines provided in B0Map, are collected in the structure

B0Map.FitOpt — Type

Data structure holding the fit parameters.

Local Fit and PHASER

n_ϕ::Int: number of golden section search (GSS) intervals for the initial phase search (with R2s == 0).

ϕ_rngs::Array: GSS intervals, associated with n_ϕ.

Δϕ2::Float64: Half interval for (optional) nonlinear optimization.

R2s_rng::Array: Search range for R2s.

ϕ_acc::Float64: Required GSS accuracy for ϕ_acc

R2s_acc::Float64: Required GSS accuracy for R2s_acc

Local Fit only

optim::Bool: Nonlinear optimiztion in addition to GSS? (Requires gradients to be implemented for the GRE model.)

PHASER only

optim_phaser::Bool: How to treat initial search in PHASER? (cf. optim for details)

balance: Max. number of data-based balancing

rapid_balance: if true, R2s = 0 will be assumed during balancing

μ_tikh::Float: (Small) Tikhonov regularization parameter

K::Vector{Int}: Fourier Kernel size

multi_scale::Bool: if true, gradient based fitting will start with smooth kernel

os_fac::Vector{Float64}: oversampling factor

redundancy::Float64:

subsampling::Symbol: subsampling strategy (:fibonacci or :random)

optim_balance::Bool: Optimization during balancing of local and gradient fit

local_fit::Bool: Perform a final local fit, based upon PHASER.

General

n_chunks::Int: Number of chunks to profit from multi-threaded execution.

accel::Symbol: Which acceleration technique should be used. Supported values are :mt (multi-threading). (:cuda not implemented yet)

Remark

Use set_num_phase_intervals to modify n_ϕ, since only then ϕ_rngs will be set properly.

To set them up, we generate an instance with default values, using the constructor

B0Map.fitOpt — Function

fitOpt(ϕ_scale = 1.0)

Default constructor for FitOpt

Argument

ϕ_scale::Float64: See set_num_phase_intervals

Remark

Typical use case:

Generate instance of FitOpt with this constructor.

Modify parameters as needed.

Example

# generate FitOpt instance with default options
fitopt = BM.fitOpt()

# modify whatever we like to modify
BM.set_num_phase_intervals(fitpar, fitopt, 4)
fitopt.ϕ_acc = 1.e-6
...

Note

To change the default phase search intervals, do not directly change the field fitopt.n_ϕ, but call the following routine:

B0Map.set_num_phase_intervals — Function

set_num_phase_intervals(fitpar, fitopt, n_ϕ)

Define, how to search minimum of χ².

Arguments

fitpar::FitPar: Fit parameters

fitopt::FitOpt: Fit options

n_ϕ: Number of search intervals

Remark

n_ϕ > 0: Split [-π, π) into n_ϕ subintervals and store them in fitopt.ϕ_rngs.

n_ϕ == 0: Clear fitopt.ϕ_rngs for a nonlinear search, starting from fitpar.ϕ