Get results

Fit estimators

Note

Estimators, calculated by local_fit! are stored in the supplied structure of type FitPar.

Derived quantities

Based upon these estimators, further model-dependent information can be calculated with the routine

B0Map.calc_par — Function

calc_par(fitpar::FitPar{T}, fitopt::FitOpt, parfun::Function, res::AbstractArray) where {T<:AbstractGREMultiEcho}

Extract model specific information and store it in array res.

Arguments

fitpar::FitPar: Fit parameters

fitopt::FitOpt: Fit options

parfun::Function: Used to call parfun(gre::T) for each location in fitpar.S, based upon the local data and fit estimates.

res::AbstractArray: Allocated space for the results (size(res) == size(fitpar.S)).

Example

ff = zeros(size(fitpar.S))                          # allocate space for the results
BM.calc_par(fitpar, fitopt, BM.fat_fraction, ff)    # do the job

Predefined maps

For convenience (and partly efficiency) some more direct calls are provided as well:

B0Map.fat_fraction_map — Function

fat_fraction_map(fitpar, fitopt)

Return the fat fraction map.

Arguments

fitpar::FitPar: Fit parameters

fitopt::FitOpt: Fit options

Remarks

size(returned map) == size(fitpar.S)

Avoids the necessity to allocate space.

Relies on calc_par.

Example

ff = BM.fat_fraction_map(fitpar, fitopt)

B0Map.freq_map — Function

freq_map(fitpar)

Return the frequency map [Hz].

Arguments

fitpar::FitPar: Fit parameters

Remarks

size(returned map) == size(fitpar.S)

Simply rescales the phase map fitpar.ϕ instead of calling calc_par.

Example

fm = BM.freq_map(fitpar)