B0Map

Documentation for B0Map.

B0Map.AbstractGREMultiEcho

B0Map.GREMultiEchoWF

B0Map.GREMultiEchoWF

B0Map.GREMultiEchoWFFW

B0Map.GREMultiEchoWFFW

B0Map.GREMultiEchoWFRW

B0Map.GREMultiEchoWFRW

B0Map.B0_map_varpro

B0Map.GSS

B0Map.GSS_fit_R2s

B0Map.GSS_fit_ϕ

B0Map.calc_A

B0Map.calc_Bb

B0Map.calc_f_c_chunk

B0Map.calc_∂Bb

B0Map.calc_∂Bb

B0Map.calc_∂f

B0Map.calc_∂∂Bb

B0Map.calc_∂∂Bb

B0Map.calc_∂∂f

B0Map.coil_sensitivities

B0Map.fatTrait

B0Map.fat_fraction

B0Map.fat_fraction

B0Map.fat_fraction

B0Map.fit_chunk_GSS

B0Map.fit_chunk_GSS

B0Map.fit_chunk_optim

B0Map.greMultiEchoWF

B0Map.greMultiEchoWFFW

B0Map.greMultiEchoWFRW

B0Map.update!

B0Map.update!

B0Map.update_A!

B0Map.update_fat_fraction!

B0Map.update_fat_fraction!

B0Map.∂Bb!

B0Map.∂Bb!

B0Map.∂∂Bb!

B0Map.∂∂Bb!

VP4Optim.A

VP4Optim.Bb!

VP4Optim.Bb!

VP4Optim.par_changed!

VP4Optim.par_changed!

VP4Optim.set_data!

VP4Optim.set_data!

VP4Optim.x_changed!

VP4Optim.x_changed!

VP4Optim.y!

VP4Optim.y!

VP4Optim.y!

VP4Optim.y_model

VP4Optim.∂Bb!

VP4Optim.∂Bb!

VP4Optim.∂∂Bb!

VP4Optim.∂∂Bb!

B0Map.AbstractGREMultiEcho — Type

AbstractGREMultiEcho{Ny,Nx,Nc,T}

Abstract supertype of multi-echo GRE sequences

Type parameters

Ny::Int: Number of acquired data == number(echoes) * number(coils)

Nx::Int: Number of variable parameters x ⊆ {ϕ, R2s, ...} (relevant for optimization)

Nc::Int: Number of linear coefficients (e.g. number of coil elements)

T::Union{Float64, ComplexF64}: acquired data type

B0Map.GREMultiEchoWF — Type

VP4Optim model

Scope

RF spoiled multi-echo GRE sequence

Water-fat tissue model

Specifics

Water and fat are constrained to have equal phase at zero echo time.

Support for multiple coil elements and coil noise covariance matrix.

The fat fraction can either be calculated to minimize the cost function $\chi^2$ for fixed ϕ and R2s (this is the default setting) or set manually.

B0Map.GREMultiEchoWF — Method

GREMultiEchoWF(::Val{Ny}, ::Val{Nx}, ::Val{Nc}, ::Val{Nt}, ts, B0, ppm_fat, ampl_fat, 
    precession, x_sym, Δt, mode, cov_mat) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.GREMultiEchoWFFW — Type

VP4Optim model

Scope

RF spoiled multi-echo GRE sequence

Water-fat tissue model

Specifics

Water and fat are not constrained to have equal phase at zero echo time.

Single coil only

B0Map.GREMultiEchoWFFW — Method

GREMultiEchoWFFW(::Val{Ny}, ::Val{Nx}, ts, B0, ppm_fat, ampl_fat, 
    precession, x_sym, Δt) where {Ny,Nx}

Auxiliary function

B0Map.GREMultiEchoWFRW — Type

VP4Optim model

Scope

RF spoiled multi-echo GRE sequence

Water-fat tissue model

Specifics

Water and fat components are described by real weights, multiplied with a common phase factor.

Single coil only.

No partial derivatives implemented

B0Map.GREMultiEchoWFRW — Method

GREMultiEchoWFRW(::Val{Ny}, ::Val{Nx}, ::Val{Nc}, ::Val{Nt}, ts, B0, ppm_fat, ampl_fat,
precession, x_sym, Δt) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.B0_map_varpro — Method

B0_map(ge_con::Function, geFW_con::Function, args, data, S, bs::BSmooth;
λ_tikh=1e-6,
n_ϕ_0=6,
R2s_rng=(0.0, 1.0),
abs_acc_R2s=1e-3,
n_chunks=8Threads.nthreads())

TBW

B0Map.GSS — Method

GSS(fun::Function, var_rng; rel_acc=1e-4, abs_acc=nothing)

Search for minimum of given function with golden section search (GSS).

Returns location of minimum with absolute accuracy abs_acc.

Arguments

fun::Function: real valued function of a single real valued argument.

var_rng::T <: Union{Vector, Tuple}: search interval boundaries, length(var_rng) == 2

rel_acc::Float64 = 1e-4: compute abs_acc = rel_acc * (var_rng[2] - var_rng[1]).

abs_acc::T <: Union{Nothing, Float64} = nothing: Overrules rel_acc, if supplied.

B0Map.GSS_fit_R2s — Method

GSS_fit_R2s(gre::AbstractGREMultiEcho, ϕ, R2s_rng; R2s_acc=1e-4)

TBW

B0Map.GSS_fit_ϕ — Method

GSS_fit_ϕ(gre::AbstractGREMultiEcho, R2s, ϕ_rng; ϕ_acc=1e-4)

TBW

B0Map.calc_A — Method

calc_A(gre::GREMultiEchoWFFW{Ny,Nx}) where {Ny,Nx}

Auxiliary method

B0Map.calc_Bb — Method

calc_Bb(gre::GREMultiEchoWFFW, A)

Auxiliary method

B0Map.calc_f_c_chunk — Method

calc_f_c_chunk(gre, cis_chunk, data, ϕ, R2s, f, c)

TBW

B0Map.calc_∂Bb — Method

calc_∂Bb(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}, sy::SVector{Nsy}) where {Ny,Nx,Nc,Nt,Nsy}

Auxiliary routine

B0Map.calc_∂Bb — Method

calc_∂Bb(gre::GREMultiEchoWFFW{Ny,Nx}, A) where {Ny,Nx}

Auxiliary method

B0Map.calc_∂f — Method

calc_∂f(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.calc_∂∂Bb — Method

calc_∂∂Bb(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}, sy::SVector{Nsy}) where {Ny,Nx,Nc,Nt,Nsy}

Auxiliary routine

B0Map.calc_∂∂Bb — Method

calc_∂∂Bb(gre::GREMultiEchoWFFW{Ny,Nx}, tA) where {Ny,Nx}

Auxiliary method

B0Map.calc_∂∂f — Method

calc_∂∂f(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.coil_sensitivities — Method

coil_sensitivities(gre::GREMultiEchoWF)

Calculates and returns the coil sensitivities.

Remarks

If the coil noise covariance matrix $\Psi$ differs from the unit matrix, the routine VP4Optim.c does not return the actual coil sensitivities.

Otherwise it does.

B0Map.fatTrait — Method

fatTrait(gre::GREMultiEchoWF)

Auxiliary routine

B0Map.fat_fraction — Method

fat_fraction(gre::GREMultiEchoWFFW)

Calculates and returns fat fraction.

B0Map.fat_fraction — Method

fat_fraction(gre::GREMultiEchoWFRW)

Returns $|r_f| / (|r_w| + |r_f|)$

B0Map.fat_fraction — Method

fat_fraction(gre::GREMultiEchoWF)

Just returns the actual value of f, whether calculated automatically or set manually.

B0Map.fit_chunk_GSS — Method

fit_chunk_GSS(gre, cis_chunk, data, ϕ_rngs, R2s_rng, ϕ_acc, R2s_acc, ϕ, R2s, χ2)

TBW

B0Map.fit_chunk_GSS — Method

fit_chunk_GSS(gre, cis_chunk, data, R2s_rng, R2s_acc, ϕ, R2s, χ2)

TBW

B0Map.fit_chunk_optim — Method

fit_chunk_optim(gre, cis_chunk, data, R2s_rng, ϕ, R2s, χ2)

TBW

B0Map.greMultiEchoWF — Function

greMultiEchoWF(ts, B0, ppm_fat, ampl_fat, precession, mode=:auto_fat; 
    x_sym=nothing, Δt=nothing, n_coils=1, cov_mat=ones(ComplexF64,1,1))

Constructor

Arguments

ts::Vector{Float64}: Echo times [ms]

B0::Float64: Main field strength [T]

ppm_fat::Vector{Float64}: Chemical shift of fat peaks

ampl_fat::Vector{Float64}: relative fat peak amplitudes (all positive with sum(ampl_fat) ≈ 1)

precession::Symbol: Orientation of precession ∈ {:clockwise, :counterclockwise}

mode::Symbol: By default (mode == :auto_fat) the fat fraction f is calculated automatically. Alternatively (mode == :manual_fat), it can also be set manually.

x_sym::Vector{Symbol}: Vector of variable parameters ∈ {:ϕ, :R2s}, default: [:ϕ, :R2s]

Δt::Union{Float64, Nothing}: Allows to adjust the frequency bandwidth $2π/Δt$ in case of non-equidistant echoes. Default: Δt equals the average echo spacing.

n_coils::Integer: Number of coil elements (default == 1). Note: n_coils == Nc

cov_mat::AbstractMatrix: Coil noise covariance matrix, if available. Default: Nc x Nc unit matrix

B0Map.greMultiEchoWFFW — Method

greMultiEchoWFFW(ts, B0, ppm_fat, ampl_fat, precession; x_sym=[:ϕ, :R2s], Δt=nothing)

Constructor

Arguments

ts::Vector{Float64}: Echo times [ms]

B0::Float64: Main field strength [T]

ppm_fat::Vector{Float64}: Chemical shift of fat peaks

ampl_fat::Vector{Float64}: relative fat peak amplitudes (all positive with sum(ampl_fat) ≈ 1)

precession::Symbol: Orientation of precession ∈ {:clockwise, :counterclockwise}

x_sym::Vector{Symbol}: Vector of variable parameters ∈ {:ϕ, :R2s}, default: [:ϕ, :R2s]

Δt::Union{Float64, Nothing}: Allows to adjust the frequency bandwidth $2π/Δt$ in case of non-equidistant echoes. Default: Δt equals the average echo spacing.

B0Map.greMultiEchoWFRW — Method

greMultiEchoWFRW(ts, B0, ppm_fat, ampl_fat, precession; x_sym=[:ϕ, :R2s], Δt=nothing)

Constructor

Arguments

ts::Vector{Float64}: Echo times [ms]

B0::Float64: Main field strength [T]

ppm_fat::Vector{Float64}: Chemical shift of fat peaks

ampl_fat::Vector{Float64}: relative fat peak amplitudes (all positive with sum(ampl_fat) ≈ 1)

precession::Symbol: Orientation of precession ∈ {:clockwise, :counterclockwise}

mode::Symbol: By default (mode == :auto_fat) the fat fraction f is calculated automatically. Alternatively (mode == :manual_fat), it can also be set manually.

x_sym::Vector{Symbol}: Vector of variable parameters ∈ {:ϕ, :R2s}, default: [:ϕ, :R2s]

Δt::Union{Float64, Nothing}: Allows to adjust the frequency bandwidth $2π/Δt$ in case of non-equidistant echoes. Default: Δt equals the average echo spacing.

B0Map.update! — Method

update!(gre::GREMultiEchoWFRW{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.update! — Method

update!(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.update_A! — Method

update_A!(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.update_fat_fraction! — Method

update_fat_fraction!(::AutoFat, gre::GREMultiEchoWF)

Auxiliary routine

B0Map.update_fat_fraction! — Method

update_fat_fraction!(::ManualFat, ::GREMultiEchoWF)

Auxiliary routine

B0Map.∂Bb! — Method

∂Bb!(::AutoFat, gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.∂Bb! — Method

∂Bb!(::ManualFat, gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.∂∂Bb! — Method

∂∂Bb!(::AutoFat, gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

B0Map.∂∂Bb! — Method

∂∂Bb!(::ManualFat, gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Auxiliary routine

VP4Optim.A — Method

VP.A(gre::GREMultiEchoWFFW)

Method described in VP4Optim

VP4Optim.Bb! — Method

VP.Bb!(gre::GREMultiEchoWFFW)

Method described in VP4Optim

VP4Optim.Bb! — Method

VP.Bb!(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}) where {Ny,Nx,Nc,Nt}

Method described in VP4Optim

VP4Optim.par_changed! — Method

VP.par_changed!(gre::GREMultiEchoWFRW)

Method described in VP4Optim TBW

VP4Optim.par_changed! — Method

VP.par_changed!(gre::GREMultiEchoWF)

Method described in VP4Optim TBW

VP4Optim.set_data! — Method

VP.set_data!(gre::GREMultiEchoWFRW{Ny,Nx,Nc,Nt}, data::AbstractArray) where {Ny,Nx,Nc,Nt}

Stores data

Arguments

data::AbstractArray: Complex data as a vector of length Ny/2.

Remarks

Input is transformed into a real vector and handed over to VP4Optim.y!.

VP4Optim.set_data! — Method

VP.set_data!(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}, data::AbstractArray) where {Ny,Nx,Nc,Nt}

Sets new data and transforms them, if necessary.

Arguments

data::AbstractArray: Complex data for all time points and coils. The elements in data must be ordered such that the expression SMatrix{Nt,Nc}(data) produces the correct data matrix $y_{j\gamma}$.

Remarks

The data are transformed based upon the coil noise covariance matrix, as described in the documentation.

This is followed by a call of VP4Optim.y! with the transformed data as arguments.

VP4Optim.x_changed! — Method

VP.x_changed!(gre::GREMultiEchoWFRW)

Method described in VP4Optim

VP4Optim.x_changed! — Method

VP.x_changed!(gre::GREMultiEchoWF)

Method described in VP4Optim

VP4Optim.y! — Method

VP.y!(gre::GREMultiEchoWFRW{Ny,Nx,Nc,Nt}, new_y::AbstractArray) where {Ny,Nx,Nc,Nt}

Method described in VP4Optim

VP4Optim.y! — Method

VP.y!(gre::GREMultiEchoWF{Ny,Nx,Nc,Nt}, new_y::AbstractArray) where {Ny,Nx,Nc,Nt}

Method described in VP4Optim

VP4Optim.y! — Method

VP.y!(gre::GREMultiEchoWFFW{Ny,Nx}, new_y::AbstractArray) where {Ny,Nx}

Method described in VP4Optim

VP4Optim.y_model — Method

VP.y_model(gre::GREMultiEchoWF)

Method described in VP4Optim

VP4Optim.∂Bb! — Method

VP.∂Bb!(gre::GREMultiEchoWFFW)

Method described in VP4Optim

VP4Optim.∂Bb! — Method

VP.∂Bb!(gre::GREMultiEchoWF)

Method described in VP4Optim

VP4Optim.∂∂Bb! — Method

VP.∂∂Bb!(gre::GREMultiEchoWF)

Method described in VP4Optim

VP4Optim.∂∂Bb! — Method

VP.∂∂Bb!(gre::GREMultiEchoWFFW{Ny,Nx}) where {Ny,Nx}

Method described in VP4Optim